(#999) Introduce trans_mat_vec and make mat3x3 header-only
