+++ src/math/point.c
@@ -111,27 +111,6 @@ float rad_to_deg(float a)
     return 180 / PI * a;
 }
 
-Point point_mat3x3_product(Point p, mat3x3 m)
-{
-    /* Convert p to Homogeneous coordinates */
-    const float homo_p[3] = {p.x, p.y, 1};
-
-    /* Transform p with matrix m */
-    const float trans_p[3] = {
-        homo_p[0] * m.M[0][0] + homo_p[1] * m.M[0][1] + homo_p[2] * m.M[0][2],
-        homo_p[0] * m.M[1][0] + homo_p[1] * m.M[1][1] + homo_p[2] * m.M[1][2],
-        homo_p[0] * m.M[2][0] + homo_p[1] * m.M[2][1] + homo_p[2] * m.M[2][2]
-    };
-
-    /* Convert p back to Cartesian coordinates */
-    const Point result_p = {
-        .x = trans_p[0] / trans_p[2],
-        .y = trans_p[1] / trans_p[2]
-    };
-
-    return result_p;
-}
-
 Vec vec_norm(Vec v)
 {
     // TODO(#657): math/point/vec_norm: using vec_length is too expensive
